[FEAT/#312] 오프보딩 로직 수정

+++ app/src/main/java/com/byeboo/app/data/mapper/quest/QuestStateMapper.kt
@@ -9,7 +9,7 @@ fun QuestCountResponseDto.toDomain(): QuestStateModel =
     QuestStateModel(
         todayComplete = this.todayComplete,
         userCurrentStatus = this.userCurrentStatus,
-        count = this.count.toLong(),
+        count = this.count,
     )
 
 fun QuestDialogueResponseDto.toDomain(): QuestDialogue =
